app/requirements.txt
Django
daphne
gunicorn
psycopg2-binary
channels
channels_redis
websockets
jsonschema
pytest
pytest-asyncio
django-environ #for env variables in settings
python-dotenv
Pillow
djangorestframework
djangorestframework-simplejwt
django-oauth-toolkit
django-filter
django-extensions
django-cors-headers